Move interceptors to the top level
